requirements.txt
Django==3.0.5
django-environ==0.4.5
gunicorn==20.0.4
whitenoise==5.0.1
djangorestframework==3.11.0